Remove platform_specific as a submodule.
The platform_specific wiil not be submodule anymore, this directory is
put into .gitignore. The platform specific dir would be included as an
independent subrepor without any formal binding to this general theme repo.

The reason is that the general theme repo cannot be bound to a specific
platform theme becouse for every platform the platform_specific is
different, while general theme is the same.